+++ src/DeepSeekCreditCheck.Tests/Services/DeepSeekApiClientTests.cs
@@ -1,4 +1,5 @@
 using System.Net;
+using System.Net.Http;
 using Moq;
 using Moq.Protected;
 using DeepSeekCreditCheck.Core.Services;